Modularize GPT Scripts (#33)
- **Refactor:** Modularize code, create new files for:
  - `handleChat`
  - `postRequest`
  - `updateSystemMessage`

* **Style:** Update message and input border colors, rename heading
